refactor(realm): WalletDataManger 대신 컨텍스트별 Repository로 분리
- 일부 메서드명 변경
