+++ lib/repository/realm/converter/multisig_wallet.dart
@@ -2,7 +2,7 @@ import 'package:coconut_wallet/model/wallet/multisig_signer.dart';
 import 'package:coconut_wallet/model/wallet/multisig_wallet_list_item.dart';
 import 'package:coconut_wallet/repository/realm/model/coconut_wallet_model.dart';
 
-MultisigWalletListItem mapRealmMultisigWalletToMultisigWalletListItem(
+MultisigWalletListItem mapRealmToMultisigWalletItem(
     RealmMultisigWallet realmMultisigWallet, String? decryptedDescriptor) {
   return MultisigWalletListItem(
       id: realmMultisigWallet.id,
